[REF] r_category: split allocation file
