+++ resource_planning/models/__init__.py
@@ -1,4 +1,5 @@
 from . import resource_location
 from . import resource
 from . import resource_allocation
+from . import resource_category
 from . import res_config
